# artifacts/api-server/src/routes/announcements.ts
import { Router, type IRouter, type NextFunction, type Request, type Response } from "express";
import multer from "multer";
import path from "path";
import fs from "fs";
import { eq, asc, sql, ne } from "drizzle-orm";
import { db, announcementsTable } from "@workspace/db";
import { mediaStore } from "../lib/storage";
import { maxUploadBytes, uploadTooLargeMessage } from "../lib/upload-limit";
import { parseFormBoolean } from "../lib/form-values";
import { parseYouTubeUrl } from "@workspace/db/youtube";
import { parseAnnouncementOrientation } from "@workspace/db/orientation";
import { normalizeDisplayText } from "../lib/slide-caption";
import {
  ListAnnouncementsResponse,
  ListAnnouncementsResponseItem,
  ListActiveAnnouncementsResponse,
  ListActiveAnnouncementsResponseItem,
  CreateAnnouncementBody,
  CreateAnnouncementResponse,
  GetAnnouncementParams,
  GetAnnouncementResponse,
  UpdateAnnouncementParams,
  UpdateAnnouncementBody,
  UpdateAnnouncementResponse,
  DeleteAnnouncementParams,
  ToggleAnnouncementParams,
  ToggleAnnouncementResponse,
  GetAnnouncementStatsResponse,
  ReorderAnnouncementsBody,
} from "@workspace/api-zod";

const router: IRouter = Router();

const uploadsDir = path.resolve(process.cwd(), "uploads");

/**
 * Raised by the multer fileFilter for non-image uploads. A dedicated class lets
 * uploadImage recognize the rejection by identity instead of matching the error
 * message string, so the two ends cannot drift apart silently.
 */
class UnsupportedImageTypeError extends Error {}

const upload = multer({
  storage: multer.memoryStorage(),
  limits: { fileSize: maxUploadBytes() },
  fileFilter: (_req, file, cb) => {
    if (file.mimetype.startsWith("image/")) cb(null, true);
    else cb(new UnsupportedImageTypeError("Only image files are allowed"));
  },
});

/**
 * Wraps multer so that its errors become explicit HTTP responses. On Vercel the
 * request body cap is lower than the default limit, so an oversized upload is an
 * expected outcome and must return a message the operator can act on.
 */
function uploadImage(req: Request, res: Response, next: NextFunction): void {
  upload.single("image")(req, res, (err: unknown) => {
    if (!err) {
      next();
      return;
    }
    if (err instanceof multer.MulterError && err.code === "LIMIT_FILE_SIZE") {
      res.status(413).json({ error: uploadTooLargeMessage(maxUploadBytes()) });
      return;
    }
    if (err instanceof UnsupportedImageTypeError) {
      res.status(400).json({ error: "Apenas arquivos de imagem são aceitos." });
      return;
    }
    next(err);
  });
}

async function persistImage(file: Express.Multer.File): Promise<string> {
  return mediaStore().put(file.buffer, file.mimetype, file.originalname);
}

/**
 * One-time migration of images written to local disk before App Storage existed.
 * Gated on PRIVATE_OBJECT_DIR: outside Replit there is no legacy disk to read,
 * and running it would issue a full table scan on every cold start.
 */
async function migrateLegacyImages() {
  const rows = await db.select().from(announcementsTable);
  for (const row of rows) {
    if (!row.imageUrl || !row.imageUrl.startsWith("/api/uploads/")) continue;
    const filename = row.imageUrl.split("/").pop();
    if (!filename) continue;
    const filePath = path.join(uploadsDir, filename);
    if (!fs.existsSync(filePath)) continue;
    try {
      const storedPath = await mediaStore().put(
        fs.readFileSync(filePath),
        "image/" + path.extname(filename).slice(1),
        filename,
      );
      await db.update(announcementsTable).set({ imageUrl: storedPath }).where(eq(announcementsTable.id, row.id));
    } catch (error) {
      console.error(`Could not migrate announcement image ${row.id}`, error);
    }
  }
}

if (process.env.PRIVATE_OBJECT_DIR) {
  void migrateLegacyImages();
}

/**
 * Biblioteca de peças: tudo menos as artes de aviso urgente, que só existem
 * para a TV registrar a exibição e não podem ser escolhidas para campanha.
 * Separada para o teste conferir o SQL sem banco.
 */
export function buildAnnouncementsListQuery() {
  return db
    .select()
    .from(announcementsTable)
    .where(ne(announcementsTable.source, "alert"))
    .orderBy(asc(announcementsTable.displayOrder), asc(announcementsTable.createdAt));
}

type YouTubeFields = {
  mediaKind: "image" | "youtube_video" | "youtube_playlist";
  youtubeId: string | null;
  playbackMode: "natural" | "capped";
  audioMode: "muted" | "sound";
};

/**
 * Deriva os campos de YouTube a partir do corpo (multipart => strings).
 * Retorna { error } quando o link é obrigatório mas inválido/ausente.
 */
function readYouTubeFields(body: Record<string, unknown>): YouTubeFields | { error: string } {
  const rawKind = body.mediaKind != null ? String(body.mediaKind) : "image";
  const mediaKind =
    rawKind === "youtube_video" || rawKind === "youtube_playlist" ? rawKind : "image";

  const playbackMode = String(body.playbackMode) === "natural" ? "natural" : "capped";
  const audioMode = String(body.audioMode) === "sound" ? "sound" : "muted";

  if (mediaKind === "image") {
    return { mediaKind: "image", youtubeId: null, playbackMode, audioMode };
  }

  const url = body.youtubeUrl != null ? String(body.youtubeUrl) : "";
  const ref = parseYouTubeUrl(url);
  if (!ref) return { error: "Link do YouTube inválido" };
  if (ref.kind !== mediaKind) {
    return { error: `O link não corresponde ao tipo selecionado (${mediaKind})` };
  }
  return { mediaKind: ref.kind, youtubeId: ref.id, playbackMode, audioMode };
}

router.get("/announcements", async (req, res): Promise<void> => {
  const rows = await buildAnnouncementsListQuery();
  res.json(ListAnnouncementsResponse.parse(rows));
});

router.post(
  "/announcements",
  uploadImage,
  async (req, res): Promise<void> => {
    const body = {
      title: req.body.title,
      displayText: req.body.displayText != null ? String(req.body.displayText) : undefined,
      showText: parseFormBoolean(req.body.showText),
      duration: req.body.duration != null ? Number(req.body.duration) : undefined,
    };
    const parsed = CreateAnnouncementBody.safeParse(body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }

    const yt = readYouTubeFields(req.body);
    if ("error" in yt) {
      res.status(400).json({ error: yt.error });
      return;
    }

    const orientation = parseAnnouncementOrientation(req.body.orientation);
    if (orientation === null) {
      res.status(400).json({ error: "Orientação inválida" });
      return;
    }

    let imageUrl: string | null = null;
    if (req.file) {
      try {
        imageUrl = await persistImage(req.file);
      } catch (error) {
        res.status(502).json({ error: "Could not persist image in object storage" });
        return;
      }
    } else if (yt.mediaKind === "image") {
      res.status(400).json({ error: "Image file is required" });
      return;
    }

    const maxOrderRow = await db
      .select({ maxOrder: sql<number>`COALESCE(MAX(${announcementsTable.displayOrder}), -1)` })
      .from(announcementsTable);
    const nextOrder = (maxOrderRow[0]?.maxOrder ?? -1) + 1;
    const [row] = await db
      .insert(announcementsTable)
      .values({
        title: parsed.data.title,
        displayText: normalizeDisplayText(parsed.data.displayText),
        showText: parsed.data.showText ?? false,
        imageUrl,
        mediaKind: yt.mediaKind,
        youtubeId: yt.youtubeId,
        playbackMode: yt.playbackMode,
        audioMode: yt.audioMode,
        // Ausente (cliente antigo): fica o default do banco, landscape.
        ...(orientation ? { orientation } : {}),
        duration: parsed.data.duration ?? 10,
        displayOrder: nextOrder,
      })
      .returning();
    res.status(201).json(CreateAnnouncementResponse.parse(row));
  }
);

router.get("/announcements/active", async (req, res): Promise<void> => {
  const rows = await db
    .select()
    .from(announcementsTable)
    .where(eq(announcementsTable.isActive, true))
    .orderBy(asc(announcementsTable.displayOrder), asc(announcementsTable.createdAt));
  res.json(ListActiveAnnouncementsResponse.parse(rows));
});

/**
 * Contagem da tela Peças: ignora as artes de aviso urgente para os números
 * baterem com a lista (que já as esconde). Separada para o teste conferir o
 * SQL sem banco.
 */
export function buildAnnouncementStatsQuery() {
  return db
    .select({
      total: sql<number>`COUNT(*)::int`,
      active: sql<number>`COUNT(*) FILTER (WHERE ${announcementsTable.isActive} = true)::int`,
      inactive: sql<number>`COUNT(*) FILTER (WHERE ${announcementsTable.isActive} = false)::int`,
    })
    .from(announcementsTable)
    .where(ne(announcementsTable.source, "alert"));
}

router.get("/announcements/stats", async (req, res): Promise<void> => {
  const [totals] = await buildAnnouncementStatsQuery();
  res.json(GetAnnouncementStatsResponse.parse(totals));
});

router.post("/announcements/reorder", async (req, res): Promise<void> => {
  const parsed = ReorderAnnouncementsBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  // Peças de painel são reordenáveis de propósito, sem checar `source` aqui.
  // Diferente de PATCH/toggle/DELETE, `displayOrder` de uma peça gerada não
  // controla nada que chega numa TV: slides de painel tocam na ordem de
  // panels.id + panel_slides.page_no (lib/panels/device-slides.ts), a
  // playlist do device segue device_playlist.displayOrder (peças de painel
  // nunca entram nessa tabela), e campanhas seguem a ordem de campaign id.
  // `displayOrder` aqui só decide a posição na lista do admin — não é parte
  // do invariante "o artefato gerado é imutável" que os outros três guardam.
  await Promise.all(
    parsed.data.ids.map((id, index) =>
      db
        .update(announcementsTable)
        .set({ displayOrder: index })
        .where(eq(announcementsTable.id, id))
    )
  );
  res.json({ ok: true });
});

router.get("/announcements/:id", async (req, res): Promise<void> => {
  const params = GetAnnouncementParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [row] = await db
    .select()
    .from(announcementsTable)
    .where(eq(announcementsTable.id, params.data.id));
  if (!row) {
    res.status(404).json({ error: "Announcement not found" });
    return;
  }
  res.json(GetAnnouncementResponse.parse(row));
});

router.patch(
  "/announcements/:id",
  uploadImage,
  async (req, res): Promise<void> => {
    const params = UpdateAnnouncementParams.safeParse(req.params);
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }
    const orientation = parseAnnouncementOrientation(req.body.orientation);
    if (orientation === null) {
      res.status(400).json({ error: "Orientação inválida" });
      return;
    }
    // FormData envia todos os campos como string — coage duration e showText antes do Zod
    const body: Record<string, unknown> = {};
    if (req.body.title !== undefined) body.title = req.body.title;
    if (req.body.displayText !== undefined) {
      body.displayText = normalizeDisplayText(String(req.body.displayText));
    }
    const showText = parseFormBoolean(req.body.showText);
    if (showText !== undefined) body.showText = showText;
    if (req.body.duration !== undefined && req.body.duration !== "") {
      body.duration = Number(req.body.duration);
    }
    const parsed = UpdateAnnouncementBody.safeParse(body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }

    const [existing] = await db
      .select()
      .from(announcementsTable)
      .where(eq(announcementsTable.id, params.data.id));
    if (!existing) {
      res.status(404).json({ error: "Announcement not found" });
      return;
    }
    if (existing.source === "alert") {
      // A arte é gerada a partir do aviso; mexer nela a desviaria do texto.
      res.status(409).json({ error: "Arte de aviso urgente não pode ser alterada." });
      return;
    }
    if (existing.source === "panel") {
      // Editar o PNG gerado quebraria a relação com o cadastro que o produziu.
      // A ação certa é despublicar o painel no portal do cliente.
      res.status(409).json({ error: "Peça gerada por painel do cliente. Edite o painel." });
      return;
    }

    const updates: Record<string, unknown> = { ...parsed.data };

    if (orientation) updates.orientation = orientation;

    // Campos de YouTube só são reprocessados se o cliente enviou mediaKind.
    if (req.body.mediaKind !== undefined) {
      const yt = readYouTubeFields(req.body);
      if ("error" in yt) {
        res.status(400).json({ error: yt.error });
        return;
      }
      updates.mediaKind = yt.mediaKind;
      updates.youtubeId = yt.youtubeId;
      updates.playbackMode = yt.playbackMode;
      updates.audioMode = yt.audioMode;
      if (yt.mediaKind !== "image" && !req.file && existing.mediaKind === "image") {
        // Trocou de imagem para YouTube sem enviar nova imagem: limpa o poster
        // antigo. Se a peça já era YouTube, o formulário reenvia mediaKind em
        // todo PATCH (mesmo sem trocar nada), e isso não pode apagar um
        // poster próprio que já estava salvo.
        updates.imageUrl = null;
      }
    }

    if (req.file) {
      let imageUrl: string;
      try {
        imageUrl = await persistImage(req.file);
      } catch (error) {
        res.status(502).json({ error: "Could not persist image in object storage" });
        return;
      }
      updates.imageUrl = imageUrl;
    }

    if (Object.keys(updates).length === 0) {
      res.json(UpdateAnnouncementResponse.parse(existing));
      return;
    }

    const [row] = await db
      .update(announcementsTable)
      .set(updates)
      .where(eq(announcementsTable.id, params.data.id))
      .returning();

    if (req.file && existing.imageUrl) {
      try {
        await mediaStore().remove(existing.imageUrl);
      } catch (error) {
        req.log.error({ err: error }, "Could not remove replaced announcement image");
      }
    }

    res.json(UpdateAnnouncementResponse.parse(row));
  }
);

router.delete("/announcements/:id", async (req, res): Promise<void> => {
  const params = DeleteAnnouncementParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [existing] = await db
    .select()
    .from(announcementsTable)
    .where(eq(announcementsTable.id, params.data.id));
  if (!existing) {
    res.status(404).json({ error: "Announcement not found" });
    return;
  }
  if (existing.source === "alert") {
    // Apagar a arte sozinha deixaria o aviso apontando para nada; o caminho
    // certo é encerrar o aviso.
    res.status(409).json({ error: "Arte de aviso urgente não pode ser apagada; encerre o aviso." });
    return;
  }
  if (existing.source === "panel") {
    // Apagar só o registro deixaria panel_slides apontando para nada, com o
    // painel ainda marcado como publicado. A ação certa é despublicar o
    // painel no portal do cliente.
    res.status(409).json({ error: "Peça gerada por painel do cliente. Edite o painel." });
    return;
  }
  const [row] = await db
    .delete(announcementsTable)
    .where(eq(announcementsTable.id, params.data.id))
    .returning();
  if (!row) {
    res.status(404).json({ error: "Announcement not found" });
    return;
  }
  if (row.imageUrl) {
    try {
      await mediaStore().remove(row.imageUrl);
    } catch (error) {
      req.log.error({ err: error }, "Could not remove deleted announcement image");
    }
  }
  res.sendStatus(204);
});

router.patch("/announcements/:id/toggle", async (req, res): Promise<void> => {
  const params = ToggleAnnouncementParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [existing] = await db
    .select()
    .from(announcementsTable)
    .where(eq(announcementsTable.id, params.data.id));
  if (!existing) {
    res.status(404).json({ error: "Announcement not found" });
    return;
  }
  if (existing.source === "alert") {
    // Ligar/desligar a arte à mão tiraria o aviso do ar sem encerrá-lo.
    res.status(409).json({ error: "Arte de aviso urgente não pode ser alterada." });
    return;
  }
  if (existing.source === "panel") {
    // Editar o PNG gerado quebraria a relação com o cadastro que o produziu.
    // A ação certa é despublicar o painel no portal do cliente.
    res.status(409).json({ error: "Peça gerada por painel do cliente. Edite o painel." });
    return;
  }
  const [row] = await db
    .update(announcementsTable)
    .set({ isActive: !existing.isActive })
    .where(eq(announcementsTable.id, params.data.id))
    .returning();
  res.json(ToggleAnnouncementResponse.parse(row));
});

export default router;