# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Digital Signage API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List all announcements
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (multipart/form-data)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "400":
          description: Bad request
...
  /announcements/reorder:
    post:
      operationId: reorderAnnouncements
      tags: [announcements]
      summary: Reorder announcements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: OK
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "404":
          description: Not found
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "404":
          description: Not found
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /announcements/{id}/toggle:
    patch:
      operationId: toggleAnnouncement
      tags: [announcements]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "404":
          description: Not found
...
  /analytics/announcements/{announcementId}:
    get:
      operationId: getAnnouncementAnalytics
      tags: [analytics]
      parameters:
        - { name: announcementId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementAnalytics"
        "404":
          description: Not found
...
components:
  schemas:
...
    Announcement:
      type: object
      required: [id, title, isActive, displayOrder, duration, showText, mediaKind, playbackMode, audioMode, orientation, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        # Texto exibido na TV; `title` é apenas o rótulo interno do painel.
        displayText: { type: string, nullable: true }
        showText: { type: boolean }
        imageUrl: { type: string, nullable: true }
        mediaKind: { type: string }
        youtubeId: { type: string, nullable: true }
        playbackMode: { type: string }
        audioMode: { type: string }
        # "landscape" | "portrait". Sem enum pelo mesmo motivo de mediaKind:
        # um valor novo não pode derrubar a lista inteira.
        orientation: { type: string }
        isActive: { type: boolean }
        displayOrder: { type: integer }
        duration: { type: integer }
        # "admin" (peça subida no painel de gestão) | "panel" (gerada pela
        # publicação de um painel do cliente — o admin não edita nem apaga
        # essas linhas, só despublica o painel de origem). Sem enum, como os
        # demais campos livres desta tabela (mediaKind, playbackMode,
        # audioMode): um terceiro valor não deve derrubar o parse da lista
        # inteira em GET /announcements.
        source: { type: string }
        createdAt: { type: string, format: date-time }
    AnnouncementInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        displayText: { type: string, nullable: true }
        showText: { type: boolean }
        duration: { type: integer, minimum: 1 }
        mediaKind: { type: string }
        youtubeUrl: { type: string }
        playbackMode: { type: string }
        audioMode: { type: string }
        orientation: { type: string, enum: [landscape, portrait] }
    AnnouncementUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        displayText: { type: string, nullable: true }
        showText: { type: boolean }
        isActive: { type: boolean }
        displayOrder: { type: integer }
        duration: { type: integer, minimum: 1 }
        mediaKind: { type: string }
        youtubeUrl: { type: string }
        playbackMode: { type: string }
        audioMode: { type: string }
        orientation: { type: string, enum: [landscape, portrait] }
...
    ReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: integer }

    # ── Segments ─────────────────────────────────────────────────────────────
...
    AnnouncementAnalytics:
      type: object
      required: [announcementId, title, totalPlays, totalDuration]
      properties:
        announcementId: { type: integer }
        title: { type: string }
        totalPlays: { type: integer }
        totalDuration: { type: integer }
        totalScans: { type: integer }
        totalUniqueScans: { type: integer }
        scanRate: { type: number }
        byCampaign:
          type: array
          items:
            type: object
            required: [campaignId, campaignName, plays, scans, scanRate]
            properties:
              campaignId: { type: integer }
              campaignName: { type: string }
              plays: { type: integer }
              scans: { type: integer }
              scanRate: { type: number }
        byDevice:
          type: array
          items:
            type: object
            required: [deviceId, deviceName, clientName, plays, totalDuration]
            properties:
              deviceId: { type: integer }
              deviceName: { type: string }
              clientName: { type: string }
              plays: { type: integer }
              totalDuration: { type: integer }
...